requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.3
beautifulsoup4==4.12.3
apscheduler==3.10.4
pytz==2024.1
lxml==5.1.0
